+++ .gitignore
@@ -172,3 +172,5 @@ cython_debug/
 .idea/shelf/Uncommitted_changes_before_Checkout_at_10_05_2023,_20_46_\[Changes\]/shelved.patch
 .idea/shelf/Uncommitted_changes_before_Checkout_at_10_05_2023,_20_46_\[Changes\]1/shelved.patch
 .idea/shelf/Uncommitted_changes_before_Checkout_at_10_05_2023__20_46__Changes_.xml
+.idea/copyright/profiles_settings.xml
+.idea/copyright/netorc.xml
