+++ api/services/service.py
@@ -6,9 +6,9 @@ from typing import List
 from fastapi import APIRouter, Depends
 from sqlmodel import Session, select
 
-from miscellaneous.addons.exceptions import APIException
-from miscellaneous.database.db import get_session
-from miscellaneous.database.tables import Service
+from core.addons.exceptions import APIException
+from core.db import get_session
+from core.db.tables import Service
 
 router = APIRouter(
     prefix="/api/services/v1",
@@ -19,12 +19,12 @@ router = APIRouter(
 @router.get("/service", response_model=List[Service])
 def get_service(id: str = None, session: Session = Depends(get_session)) -> list:
     """
-    Returns a list of service/'s from the database.
-    Arguments can be included in the uri to be added to the database query.
+    Returns a list of service/'s from the db.
+    Arguments can be included in the uri to be added to the db query.
 
     Args:
         id: an optional service id filter.
-        session: database session dependency.
+        session: db session dependency.
 
     Returns:
         A list of dictionary/s containing attributes for a service.
@@ -50,7 +50,7 @@ def get_service(id: str = None, session: Session = Depends(get_session)) -> list
 @router.post("/service", response_model=List[Service])
 async def post_service(service: Service) -> list:
     """
-    Creates new service/s. The service/s is added to the database once the task completes.
+    Creates new service/s. The service/s is added to the db once the task completes.
 
     Args:
 
@@ -64,7 +64,7 @@ async def post_service(service: Service) -> list:
 @router.patch("/service/{id}", response_model=Service)
 async def patch_service(id: int) -> list:
     """
-    Amend a service. The service is updated in the database once the task completes.
+    Amend a service. The service is updated in the db once the task completes.
 
     Args:
 
@@ -78,7 +78,7 @@ async def patch_service(id: int) -> list:
 @router.delete("/service/{id}", response_model=Service)
 async def delete_service(id: int) -> list:
     """
-    Remove a service. The service is deleted in the database once the task completes.
+    Remove a service. The service is deleted in the db once the task completes.
 
     Args:
