+++ core/addons/celery.py
