+++ core/db/__init__.py
@@ -1,0 +1,10 @@
+from sqlmodel import create_engine, Session
+
+from settings import settings
+
+engine = create_engine(settings.database, echo=True)
+
+
+def get_session():
+    with Session(engine) as session:
+        yield session
